azuredevops/python-samples/requirements.txt
azure-devops==5.0.0b5
